# Actually use the solver when starting rewriting