When processing an execute request, we start z3 in `JsonRpc.hs` and pass it to `performRewrite`. However, we then actually start rewriting without a solver. This PR makes sire the initial state in `performRewrite` actually is initialized with a solver.